# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/promotion": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Lista Promociones registradas. Valida autorizacion de usuario",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Recupera un listado de las Promociones registrados",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "used",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Promotion"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Las promociones se crean con Usado en falso Valida autorizacion de usuario; que el codigo de promocion no este registrado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Crea registro de promociones de los",
                "parameters": [
                    {
                        "description": "create promotion",
                        "name": "PromoDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PromotionDTOCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/promotion/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Obtiene informacion de la Promocion segun su ID Valida autorizacion de usuario; existencia de promocion a mostrar",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Recupera una Promocion segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Promotion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promotion"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "modifica la promocion segun su ID Valida autorizacion de usuario; que exista la promocion a editar; que el codigo de promocion no este registrado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Modifica una Promocion segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Promotion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update promotion",
                        "name": "PromoDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PromotionDTOUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promotion"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Elimina una promocion segun su ID Valida Autorizacion de usuario; promocion a eliminar exista; promocion asociada a una orden",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Elimina una promocion segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Promotion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.MessageInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PromotionDTOCreate": {
            "type": "object",
            "required": [
                "code",
                "discount",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "discount": {
                    "type": "number",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "dto.PromotionDTOUpdate": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "used": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Promotion": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "used": {
                    "type": "boolean"
                }
            }
        },
...
    }
}